Disable UNK alerts by default
UNK alerts are generated due to Amazon's throttling limits, this causes a high number of alerts which are false only because the Cloudwatch API responses during throttling. As a workaround this should be disabled by default, with an option to enable explicitly since the alerts are not actionable (unless limits are raised by Amazon) and since the alerts clear on the next OK response. This causes false alerts 99% of the time.