Use = for [ ... ] tests instead of bash's ==
For Bash, the builtin command `test` resp. `[` accepts both `==` as well
as `=` as an operator. POSIX however only specifies `=`. Thus, if the
scripts are used with a `/bin/sh` which is not Bash, e.g. dash on Debian
derivates, the test fail with:

[: 1: unexpected operator

This commit ensures that only the POSIX-conformiing operator `=` is
used. For Bash, it doesn't make any difference, for other shells, it
ensures that the scripts still work.